Title: 0에 가장 가까운 합, Time: 347ms, Memory: 21MB, Status: Wrong Answer - CodeTree

+++ 240807/0에 가장 가까운 합/sum-closest-to-zero.java
@@ -5,16 +5,16 @@ public class Main {
     public static void main(String[] args) throws Exception {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
-        int[] arr = new int[n+1];
+        long[] arr = new long[n+1];
         StringTokenizer st = new StringTokenizer(br.readLine());
         for(int i = 1; i <= n; i++) {
-            arr[i] = Integer.parseInt(st.nextToken());
+            arr[i] = Long.parseLong(st.nextToken());
         }
 
         Arrays.sort(arr, 1, n+1);
 
         int j = n;
-        int minValue = Integer.MAX_VALUE;
+        long minValue = Integer.MAX_VALUE;
         for(int i = 1; i <= n; i++) {
             
             while(j > i && Math.abs(arr[i] + arr[j]) <= minValue) {
